Allow to configure EctoLogger through :telemetry
For complex umbrella apps, having a global config for the :spandex_ecto
library can be limiting - particularly when a different tracer is to be
used depending on the repo.

This change allows to pass configuration to the underlying EctoLogger
module, like the `:tracer` option, through the `config` argument in
`:telemetry.attach/4`.